% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{print.tdigest}
\alias{print.tdigest}
\title{Printer for t-idgest objects}
\usage{
\method{print}{tdigest}(x, ...)
}
\arguments{
\item{x}{t-tigest object}

\item{...}{unused}
}
\description{
Printer for t-idgest objects
}
\keyword{internal}
