% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{td_add}
\alias{td_add}
\title{Add a value to the t-Digest with the specified count}
\usage{
td_add(td, val, count)
}
\arguments{
\item{td}{t-Digest object}

\item{val}{value}

\item{count}{count}
}
\value{
the original, updated \code{tdigest} object
}
\description{
Add a value to the t-Digest with the specified count
}
\examples{
td <- td_create(10)
td_add(td, 0, 1)
}
