% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{td_merge}
\alias{td_merge}
\title{Merge one t-Digest into another}
\usage{
td_merge(from, into)
}
\arguments{
\item{from, into}{t-Digests}
}
\value{
\code{into}

a \code{tdigest} object
}
\description{
Merge one t-Digest into another
}
