% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{td_quantile_of}
\alias{td_quantile_of}
\title{Return the quantile of the value}
\usage{
td_quantile_of(td, val)
}
\arguments{
\item{td}{t-Digest object}

\item{val}{value}
}
\value{
the computed quantile (\code{double})
}
\description{
Return the quantile of the value
}
