% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{td_add}
\alias{td_add}
\title{Add a value to the t-digest with the specified count}
\usage{
td_add(td, val, count)
}
\arguments{
\item{td}{t-digest object}

\item{val}{value}

\item{count}{count}
}
\description{
Add a value to the t-digest with the specified count
}
