% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{td_merge}
\alias{td_merge}
\title{Merge one t-digest into another}
\usage{
td_merge(from, into)
}
\arguments{
\item{from, into}{t-digests}
}
\value{
\code{into}
}
\description{
Merge one t-digest into another
}
