% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{td_value_at}
\alias{td_value_at}
\alias{[.tdigest}
\title{Return the value at the specified quantile}
\usage{
td_value_at(td, q)

\method{[}{tdigest}(x, i, ...)
}
\arguments{
\item{td}{t-Digest object}

\item{q}{quantile (range 0:1)}

\item{x}{a tdigest object}

\item{i}{quantile (range 0:1)}

\item{...}{unused}
}
\value{
the computed quantile (\code{double})
}
\description{
Return the value at the specified quantile
}
\examples{
td <- td_create(10)

td_add(td, 0, 1) \%>\%
  td_add(10, 1)

td_value_at(td, 0.1)
td_value_at(td, 0.5)
td[0.1]
td[0.5]
}
