% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdigest-package.R
\docType{package}
\name{tdigest}
\alias{tdigest}
\alias{tdigest-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/tdigest}
\item BugReports: \url{https://gitlab.com/hrbrmstr/tdigest/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
