% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{tquantile}
\alias{tquantile}
\title{Create a new t-digest histogram from a vector}
\usage{
tquantile(td, probs)
}
\arguments{
\item{td}{t-digest object}

\item{probs}{numeric vector of probabilities with values in range 0:1}
}
\description{
Create a new t-digest histogram from a vector
}
